requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Filters
django-filter==23.5

# API Documentation
drf-yasg==1.21.7

# Task Queue
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# Environment Variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
Pillow==10.2.0

# Reports
reportlab==4.0.9
openpyxl==3.1.2

# WSGI Server
gunicorn==21.2.0

# Development
django-extensions==3.2.3
